Only delete new variables in AssignAttributes
Also refactors component.
