+++ src/components/AssignAttributes/withAttributeHandlers.js
@@ -3,7 +3,6 @@ import { connect } from 'react-redux';
 import { formValueSelector } from 'redux-form';
 import { withHandlers, compose } from 'recompose';
 import { getVariablesForSubject } from '../../selectors/codebook';
-import { actionCreators as codebookActions } from '../../ducks/modules/protocol/codebook';
 
 const store = connect(
   (state, { entity, type, form, field }) => {
@@ -19,17 +18,13 @@ const store = connect(
       options,
     };
   },
-  { deleteVariable: codebookActions.deleteVariable },
 );
 
 const handlers = withHandlers({
-  handleDelete: ({ fields, entity, type, variable, deleteVariable }) =>
-    (index) => {
-      fields.remove(index);
-      if (variable) {
-        deleteVariable(entity, type, variable);
-      }
-    },
+  handleDelete: ({ onDelete, variable, index }) =>
+    () => onDelete(index, variable),
+  handleCreateNew: ({ index, onCreateNew }) =>
+    () => onCreateNew(index),
 });
 
 export default compose(
